Restructuring code and tests in subfolders
