+++ ragxiv/datamodel/__init__.py
@@ -1,1 +1,8 @@
-from .datamodel import ArxivPaper, Author, Method, Simulation
+# ############################################################################
+# This sub-folder contains the data model classes to be populated
+# by the LLM. These classes are very use-case specific and can thus
+# be modified to fit the needs of the application.
+# ############################################################################
+
+from .arxiv import ArxivPaper, Author
+from .dmft import Method, Simulation
