+++ ragxiv/parsing/__init__.py
@@ -1,0 +1,5 @@
+# ############################################################################
+# This sub-folder contains the parsing classes and functions
+# for the extracted text and structured data model into a
+# database.
+# ############################################################################
