+++ ragxiv/rag/__init__.py
@@ -1,0 +1,9 @@
+# ############################################################################
+# This sub-folder contains the retrieval and generation classes
+# and functions to process the text. These are used to retrieve
+# relevant chunks of text and generate structured output using
+# an LLM.
+# ############################################################################
+
+from .generator import LLMGenerator, answer_to_dict
+from .retriever import CustomRetriever, LangChainRetriever
