+++ ragxiv/text/__init__.py
@@ -1,0 +1,8 @@
+# ############################################################################
+# This sub-folder contains the text extraction and chunking classes
+# and functions to process the text. These are used to extract text
+# and chunk it into smaller parts for processing in the LLM.
+# ############################################################################
+
+from .arxiv_extractor import ArxivFetcher, TextExtractor, arxiv_fetch_and_extract
+from .chunker import Chunker
